Use 'rvslots' when fetching revisions on MW 1.32+
The only places that use 'rvprop' parameter directly are in
APISite.loadrevisions and APISite.preloadpages. Modify them to
add rvslots to the query if the MW version is >= 1.32.

The recieved pagedata format will also change. Adapt the api._update_revisions
method to the new format.

This is a temporary workaround to silence the API warnings. We perhaps need to
update the Revision class for other slot types in the future.

Bug: T200955
Change-Id: Ia4d655a46bf83a969b907280c7a9e24e8782b9a8