CosmeticChangesToolkit.removeEmptySections: Do not skip interwikis
This part of code was not working correctly. The 'interwiki' regex
skips the trailing newlines and therefore may potentially mess up
page sections. Removing this part of code did not make any test fail
and seems harmless and could only potentially lead to not removing
an empty section which is better than blanking a section by mistake.

Also note that skip_regexes[1] contains 'category' regex and there
seemed to be a typo in that part of code.

None of the tests failed after this change.

Add a test for the fixed bug.

Do minor cleanups in textlib.py:
  - _ignore_case:: No need to ignore-case a letter if it is not
    case-sensitive.
  - 'interwiki' regex patter, replace '[\s]*' with '\s*'.

Change-Id: I22b4ad92c7058ca65c6e3e6a680586dc1c2490a0