Introduce a new APISite method: mwversion
APISite.version() is widely used method. In most of its usages we
convert the version string to a MediaWikiVersion object. This makes
the code ugly and inefficient (because the version string needs to be
parsed every time).

To solve this issue introduce a new method, mwversion, that returns
MediaWikiVersion object and caches the result.

Use the new method throughout the codebase.

Change-Id: I75de443b24d2eb478724c416db8f2d734b6bb806