Remove cryptography from requirements
We only needed cryptography (and other security extras of requests)
to silence the InsecurePlatformWarning on older python versions. (T103769)

Apparently that approach may lead to a false sense of security
if the required versions are pinned to specific older ones. (T203435)

It's better to let users get that warning than to have an insecure
dependency in our requirements.

- Remove cryptography and other requests security extras
  from requirement files and setup.py.
- Use Xenial instead of trusting on python 2.7_with_system_site_packages
  Travis-CI job to avoid the related errors.

Note that although Toolforge currently has an old python version (2.7.6),
but the ssl module of the default python there is recent enough to not
have this issue. Therefore, this change will not affect Toolforge users.

Bug: T203435
Change-Id: Ia7a1672cadf45a15a12969598f95f5cc5384995d