Remove cryptography<2.3 from python<=2.7.6 requirements
We added cryptography and other requests security extras
to silence the InsecurePlatformWarning on older python versions. (T103769)
That approach may lead to a false sense of security if the required
versions are pinned to specific older ones. (T203435)

It's better to let users deal with the error rather than to have an
insecure dependency in requirements.

- Remove cryptography>=1.3.4,<2.3 from requirements
- Use requests[security] instead of manually installing each requirement.

The change will cause test failures on Travis-CI's
Python 2.7_with_system_site_packages job. Reconfigure that job to use a
Xenial dist instead of Trusty which has a more up-to-date ssl module.

Note that although Toolforge currently has an old Python version (2.7.6),
but the ssl module is recent enough to not have this issue. So, this change
will not affect Toolforge users.

Bug: T203435
Change-Id: Ia7a1672cadf45a15a12969598f95f5cc5384995d