Create a console script entry point for pywikibot
- Define the `pywikibot` entry point in setup.py.
- Move the pwb.py, generate_family_file.py, and generate_user_files.py
  file into pywikibot/_scripts directory. Redirect the invocations to
  of the old script to the new ones for backward compatibility.
  Do some cleanup and fix the path operations in the moved files.

The _scripts directory will be distributed with pywikibot installation,
making it possible for library users to generate family or user files
using `pywikibot generate_family_file` or `pywikibot generate_user_files`
commands respectively.

In the future we can use the _scripts directory as a namespace package,[1]
so that users will be able to install third-party scripts into it
using pip.

[1]:
https://packaging.python.org/guides/creating-and-discovering-plugins/#using-namespace-packages

Bug: T107629

Change-Id: I81f71d871ccfaa0f9eef4f27407e51833a9aadca