config2.py: Warn instead of RuntimeError when no user-config is found
I was trying to include the generate_user_files and generate_family_file
in the distributed PYPI package,[1] but doing so turned out to be very
hard since it's impossible to run them as part of pywikibot without
them raising RuntimeError, because pywikibot.__init__ always loads before
any such script and prevents loading them.

There is no need for the RuntimeError, we can just warn the user about
the situation and let the scripts load normally. Sounds like harmless
change.
This patch does that and will hopefully make the import system
of config2.py less confusing, and may also help with the future development
of a patch for T107629.

Change-Id: I191d7db1f473fe4f5943aa5f0ca8f582445c4e03